Add Message fix
